Nesting urldecode with html_entity_decode to make certain links work in administration/customer/
When using i.e. Swedish characters in the Title-column of the customer-table and when having "Prettify links" enabled, some links break when trying to reach
the customer detail-page from  ../administration/customers/ from what seems to be some htmlentities function or such which causes the SQL-query to fail because the 'title' cant be found.
Nesting urldecode within html_entity_decode() fixes the issue.